Separate file for JsonRpcError
